requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.2
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
factory-boy==3.2.1
Faker==15.1.1
flake8==5.0.4
flower==1.2.0
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
mccabe==0.7.0
msgpack==1.0.4
packaging==21.3
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
toposort==1.7
tornado==6.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.5.0
